+++ src/components/changeProfile/VerifyPassword.tsx
@@ -1,78 +1,36 @@
-import { FormProvider, useForm } from 'react-hook-form';
-import { z as zod } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
+import { FormProvider } from 'react-hook-form';
 import Button from '../common/Button';
-import { useVerifyPasswordMutation } from '@/api/auth/changeProfile/verifyPassword/verifyPassword.hooks';
 import AuthInput from '../auth/AuthInput';
-import { ApiErrorResponseDto } from '@/types/apiErrorType';
-import { useToast } from '@/hooks/useToast';
-import axios from 'axios';
-import { VerifyPasswordRequestParams } from '@/api/auth/changeProfile/verifyPassword/verifyPasswordType';
-
-const verifyPasswordSchema = zod.object({
-  password: zod.string().min(1, { message: '비밀번호를 입력해주세요.' }),
-});
+import { useVerifyPasswordForm } from '@/hooks/changeProfile/useVerifyPasswordForm';
 
 type VerifyPasswordProps = {
   onUserVerifyPassword: () => void;
 };
 
 const VerifyPassword = ({ onUserVerifyPassword }: VerifyPasswordProps) => {
-  const { showToast } = useToast();
-
-  const form = useForm<VerifyPasswordRequestParams>({
-    resolver: zodResolver(verifyPasswordSchema),
-    defaultValues: {
-      password: '',
-    },
-    mode: 'onChange',
-  });
-
   const {
+    form,
     register,
-    handleSubmit,
     formState: { errors },
-  } = form;
-
-  const { mutate: verifyPasswordMutation, isPending } =
-    useVerifyPasswordMutation({
-      onSuccess: () => {
-        onUserVerifyPassword();
-      },
-      onError: (error) => {
-        console.error('Login Error:', error);
-
-        if (axios.isAxiosError(error)) {
-          console.error('Axios Error Details:', {
-            response: error.response?.data,
-            status: error.response?.status,
-            headers: error.response?.headers,
-          });
-        }
-
-        const apiError = error as ApiErrorResponseDto;
-        const errorMessage =
-          apiError?.response?.data?.message ||
-          '비밀번호 확인에 실패하였습니다. 다시 시도해주세요.';
-        showToast(errorMessage);
-      },
-    });
+    handleVerifypassword,
+    isPending,
+  } = useVerifyPasswordForm();
 
   const onSubmit = () => {
-    const data = form.getValues();
-    verifyPasswordMutation(data);
+    handleVerifypassword();
+    onUserVerifyPassword();
   };
 
   return (
     <FormProvider {...form}>
       <form
-        onSubmit={handleSubmit(onSubmit)}
+        onSubmit={form.handleSubmit(onSubmit)}
         className='flex flex-col items-center gap-4 px-4 pt-[54px]'
       >
         <span className='text-[15px] font-medium text-textMainColor'>
           현재 비밀번호를 입력해주세요.
         </span>
-        <div className='flex flex-col w-full gap-2'>
+        <div className='flex w-full flex-col gap-2'>
           <AuthInput
             type='password'
             placeholder='비밀번호를 입력해주세요.'
